Add files via upload
## Working in Power BI is always fun and challenging, every time and everyday you learn new things. Its just not about creating visuals, it’s about transforming raw data into meaningful visuals to solve any business challenge.

This is my first time participating in any challenge. Working on this project was fun and it made me brush up my skills in and taught me so much. It’s rightly said, the best way to learn is to work on a project.

First a brief about the problem statement which is of a supply chain business, they wants to understand and improve their process so they can retain customers and expand to more metro cities.

You will find detail information below.

https://codebasics.io/event/codebasics-resume-project-challenge

My approach to the problem was understanding in depth the problem statement and then understand the data provide.

Clean and build a model and create measures and visuals.

The most important part is to visualize the requirements at first in your mind and draw the rough sketch of it.

I will summarize my learnings briefly, but the list is never exhaustive,

	Use of bookmarks intensively to create interactive switch buttons and hidden filter box/ information pane.
	Use of custom charts.
	Understanding of supply chain domain matrices and KPI parameters.
	Converting business requirements into visuals in turn generating insights and recommendations to point out areas of improvements.
I have tried my best to optimize the report to make it as interactive and informative

INSIGHTS

OVERALL – ELITE MART HAS LEAST INFULL% RATE OF 37.94% AND EXPRESSION STORES HAS THE HIGHEST WITH 60.83%

LOTUS MART HAS LOWEST ONTIME % WITH 28.11% COOLBLUE HAS 13.75% OF ON TIME IN FULL RATE MAKING IT THE LOWEST

COOLBLUE HAS LOWEST LIFR AND VOFR RATE ASWELL

THE NUMBER OF ORDERS ARE ALMOST EQUALLY PLACED FROM 3 OPERATING CITIES OVER THE MONTHS

VOLUME OF ORDER IS A BIT LOW OF SURAT THAN OTHER TWO CITIES

DAIRY IS THE MOST ORDERED PRODUCT TYPE.

AVERAGE LIFR IS AROUND 65.96% WITH VARIATION BETWEEN 51.53% TO 75.62%

VOFR IS IN AT 96.59% AVERAGE OVER MONTHS VARYING BETWEEN 95.08% TO 97.70%

VIJAY STORES HAVE PLACED MAXIMUM ORDERS BY VOLUME

LOTUS MART HAVE PLACE MAXIMUM NUMBER OF ORDERS

SO FAR, A TOTAL OF 31729 ORDERS HAVE BEEN RECEIVED AND 96.59% ITEMS HAVE BEEN SUCCESSFULLY DELIVERED

VADODRA HAS HIGHEST UNDELIVRED RATE OF 3.63%

RECOMMENDATION

AM MILK 250 IS THE HIGHEST ORDERED PRODCUT WITH 9.53% AND AM MILK 100 AND AM MILK 500 IS THE CONSECUTIVE HIGHEST ORDERED ITEMS WITH 3.29-3.39% UNDELIVERED RATE.

CONCENTRATING ON THIS HIGH VOLUME PRODUCT WILL HELP TO IMPROVE OVERALL FIGURES.